Fixed missing conversion

+++ src/lex/colt_lexer.cpp
@@ -23,11 +23,11 @@ namespace clt::lng
     while (auto nl = static_cast<const char*>(memchr(&text[start], '\n', size - start)))
     {
       i64 nl_index = nl - text;
-      buffer.push_back(StringView{ front + start, nl_index - start });
+      buffer.push_back(StringView{ front + start, static_cast<size_t>(nl_index - start) });
       start = nl_index + 1;
     }
     // The last line ends at the end of the file.
-    buffer.push_back(StringView{ front + start, size - start });
+    buffer.push_back(StringView{ front + start, static_cast<size_t>(size - start) });
 
     // Add an extra blank line so that we never need to handle the special case
     // of being on the last line inside the lexer and needing to not increment
